Move to common-byok
