+++ embabel-agent-common/pom.xml
@@ -20,6 +20,7 @@
     </scm>
 
     <modules>
+        <module>embabel-agent-byok</module>
         <module>embabel-agent-ai</module>
         <module>embabel-agent-webmvc</module>
     </modules>
