+++ embabel-agent-dependencies/pom.xml
@@ -44,6 +44,12 @@
 
             <dependency>
                 <groupId>com.embabel.agent</groupId>
+                <artifactId>embabel-agent-byok</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>com.embabel.agent</groupId>
                 <artifactId>embabel-agent-ai</artifactId>
                 <version>${project.version}</version>
             </dependency>
